refactor: mockito 이용하여 service 테스트하도록 수정
