--- src/main/java/wooteco/subway/dao/JdbcLineDao.java
@@ -1,86 +1,0 @@
-package wooteco.subway.dao;
-
-import java.sql.PreparedStatement;
-import java.util.List;
-import java.util.Optional;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Line;
-
-@Repository
-public class JdbcLineDao implements LineDao {
-
-    private static final RowMapper<Line> LINE_ROW_MAPPER = (resultSet, rowNum) -> {
-        return new Line(
-                resultSet.getLong("id"),
-                resultSet.getString("name"),
-                resultSet.getString("color")
-        );
-    };
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public JdbcLineDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Override
-    public long save(final Line line) {
-        final String sql = "insert into LINE (name, color) values (?, ?)";
-
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement preparedStatement = connection.prepareStatement(sql, new String[]{"id"});
-            preparedStatement.setString(1, line.getName());
-            preparedStatement.setString(2, line.getColor());
-            return preparedStatement;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
-    }
-
-    @Override
-    public boolean existLineById(final Long id) {
-        final String sql = "select exists (select * from LINE where id = ?)";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, id);
-    }
-
-    @Override
-    public boolean existLineByName(final String name) {
-        final String sql = "select exists (select * from LINE where name = ?)";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, name);
-    }
-
-    @Override
-    public boolean existLineByColor(final String color) {
-        final String sql = "select exists (select * from LINE where color = ?)";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, color);
-    }
-
-    @Override
-    public List<Line> findAll() {
-        final String sql = "select id, name, color from LINE";
-        return jdbcTemplate.query(sql, LINE_ROW_MAPPER);
-    }
-
-    @Override
-    public Optional<Line> find(final Long id) {
-        final String sql = "select id, name, color from LINE where id = ?";
-        return Optional.ofNullable(jdbcTemplate.queryForObject(sql, LINE_ROW_MAPPER, id));
-    }
-
-    @Override
-    public void update(final long id, final Line line) {
-        final String sql = "update LINE set name = ?, color = ? where id = ?";
-        jdbcTemplate.update(sql, line.getName(), line.getColor(), id);
-    }
-
-    @Override
-    public void delete(final Long id) {
-        final String sql = "delete from LINE where id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-}
