--- src/main/java/wooteco/subway/dao/JdbcStationDao.java
@@ -1,65 +1,0 @@
-package wooteco.subway.dao;
-
-import java.sql.PreparedStatement;
-import java.util.List;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Station;
-
-@Repository
-public class JdbcStationDao implements StationDao {
-
-    private static final RowMapper<Station> STATION_ROW_MAPPER = (resultSet, rowNum) -> {
-        return new Station(
-                resultSet.getLong("id"),
-                resultSet.getString("name")
-        );
-    };
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public JdbcStationDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Override
-    public long save(final Station station) {
-        final String sql = "insert into STATION (name) values (?)";
-
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement preparedStatement = connection.prepareStatement(sql, new String[]{"id"});
-            preparedStatement.setString(1, station.getName());
-            return preparedStatement;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
-    }
-
-    @Override
-    public boolean existStationById(final Long id) {
-        final String sql = "select exists (select * from STATION where id = ?)";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, id);
-    }
-
-    @Override
-    public boolean existStationByName(final String name) {
-        final String sql = "select exists (select * from STATION where name = ?)";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, name);
-    }
-
-    @Override
-    public List<Station> findAll() {
-        final String sql = "select id, name from STATION";
-        return jdbcTemplate.query(sql, STATION_ROW_MAPPER);
-    }
-
-    @Override
-    public void delete(final Long id) {
-        final String sql = "delete from STATION where id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-}
