+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,17 +1,60 @@
 package wooteco.subway.dao;
 
+import java.sql.PreparedStatement;
 import java.util.List;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
 import wooteco.subway.domain.Station;
 
-public interface StationDao {
+@Repository
+public class StationDao {
 
-    long save(Station station);
+    private static final RowMapper<Station> STATION_ROW_MAPPER = (resultSet, rowNum) -> {
+        return new Station(
+                resultSet.getLong("id"),
+                resultSet.getString("name")
+        );
+    };
 
-    boolean existStationById(Long id);
+    private final JdbcTemplate jdbcTemplate;
 
-    boolean existStationByName(String name);
+    public StationDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
 
-    List<Station> findAll();
+    public long save(final Station station) {
+        final String sql = "insert into STATION (name) values (?)";
 
-    void delete(Long id);
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement preparedStatement = connection.prepareStatement(sql, new String[]{"id"});
+            preparedStatement.setString(1, station.getName());
+            return preparedStatement;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    public boolean existStationById(final Long id) {
+        final String sql = "select exists (select * from STATION where id = ?)";
+        return jdbcTemplate.queryForObject(sql, Boolean.class, id);
+    }
+
+    public boolean existStationByName(final String name) {
+        final String sql = "select exists (select * from STATION where name = ?)";
+        return jdbcTemplate.queryForObject(sql, Boolean.class, name);
+    }
+
+    public List<Station> findAll() {
+        final String sql = "select id, name from STATION";
+        return jdbcTemplate.query(sql, STATION_ROW_MAPPER);
+    }
+
+    public void delete(final Long id) {
+        final String sql = "delete from STATION where id = ?";
+        jdbcTemplate.update(sql, id);
+    }
 }
